Move TreeItemData type out of RPC since it's used in many different places
